Check NugetRestore flag before writing message
If nuget.exe isn't in the Tools folder on the agent, we *always* emit a
warning that we can't find it and that package restore won't be
performed.  We emit this warning even if the MSBuild/VSBuild task has
the nuget restore option unchecked.  So now we'll only emit the warning
if we can't find nuget.exe *and* the option is checked.