Windows ta install bug
Somehow this checkin didnt go in. The fix was suggested by Pavan.

"If we specify –ComputerName and –Credentials, then PS is creating a new
session with the provided credentials to the given host.
This is nothing but remote execution of given command. In our case, it
is execution of TestAgent.exe remotely.
As we have observed, Win7 still has issue with remote execution of .NET.
So, I thought of removing the above options so that PS executes in the
current session" -- Pavan.
Tested the fix on windows 7 machine