1) Updated waring strings and 2) removed foreground color from verbose output (bug).
If the argument passed to the ArgumentList param is null, Powershell omits the arg altogether and caused the string.format to error. I'm unsure if this will ever be null, so to be on the safe side I enclosed it in double quotes to coalesce null with empty string before passing to the string formatter.