Convert the server name to lowercase automatically
SQL Azure only allows lowercase server names.  Today, this task will fail if there are any non-lowercase letters in the server name.  A common scenario is to deploy a SQL Azure server using an ARM Template, and name the server the same as the Resource Group, but doing a ToLower() inside the ARM template. If you are trying to create a TFS build that deploys the ARM Template then deploys a DACPAC into the new SQL Server, you likely only have the Resource Group Name as a build variable, and no convenient way to convert it to lowercase from within the build.  If this task would automatically do a ToLower on it as proposed it would solve this problem.