Fixed problem setting deployment Id
It appears the intention was to use the Commit Id when running this task from a build, and the Release URI when running from a release.  The problem is the logic was a little off.  It was assuming the build.sourceVersion would be empty when running from a release, but that's not true (at least not if you have a release definition that is linked to a build).

Changed it to first check if releaseUri is set and if so use that for deploymentId (this should not be set in a build context), and if not then use the build.sourceVersion as the deployment context.

Related to Issue #1311