Adding Cmd demand to NugetPackager task
Without this demand users are trying to use this task on Mac and Linux agents. This will silently fail.  By adding the demand you will get a error when you attempt to save your build definition if a Windows agent is not in the select pool.  This might be a good task to re-write as a Node.js task so it can be used on all agents.