Support automatic diff versioning for Service Fabric apps
I added an option to only update the packages that have changed. This is largely made possible by the 'determinstic' flag introduced in VS "15" and VS2015U2 which compiles code to the same output every time it has the same input. The user can compare against the last successful build or a specific build number.

The logs could be very helpful so I added an option to 'Log all changes' and made that the default. If not checked, we will break on first change and thus performance will be better