Make parallel build checkbox label more clear
"Maximum CPU Count" makes it sound like I get to set that value (a number!). But in fact, this is just a checkbox, and it's an indicator of whether to build in parallel or not.
The `/m` switch allows specifying a value. But since this build task doesn't accept a value, renaming it to "Build in parallel" (IMO) much better helps people know what the checkbox means without having to study the long documentary explanation.