Fix cert import issue in ServiceFabricDeploy task (#3205)
A customer was encountering an issue with the ServiceFabricDeploy task where an Access Denied error would occur when attempting to import their certificate used to connect to the cluster.

I determined that the certificate import logic was attempting to store the private key on the MachineKeys folder of the agent. But the user being run as for the task does not have permission to that folder.

The fix is to explicitly set the key storage flags parameter to use UserKeySet which will store the private key in a user-scoped folder which the user will have permission to. And I verified that the rest of the code which uses the certificate operates correctly when it is stored in this manner.