Support CompressPackage option in ServiceFabricDeploy task (#3350)
A new feature in Service Fabric SDK adds a CompressPackage option parameter to the Copy-ServiceFabricApplicationPackage cmdlet. When enabled, this parameter compresses the package prior to copying it to the image store to improve performance. The SF SDK updated the Publish-NewServiceFabricApplication and Publish-UpgradedServiceFabricApplication scripts to include this new parameter as well. We should allow users of the Service Fabric Deploy task to make use of this feature.

Two changes that are made to support this feature:

Referenced publish profiles can have the CompressPackage option specified. This setting is now passed along to the call to Service Fabric SDK's Publish-NewServiceFabricApplication and Publish-UpgradedServiceFabricApplication scripts.
Uses have the option to specify the CompressPackage option as part of the task itself which, if it's enabled, will override the setting in the publish profile.