Modified bug fix in AzurePowerShell.ps1
Modified per advice from reviewer niadak. Reverted previous change and declared the variable $__vstsAzPSInlineScriptPath at line 28. This avoids the "cannot be retrieved" error occurring in the finally block in the case where the called .ps1 script has executed "Set-PsDebug -Strict". (The potential alternate fix, testing $scriptType in the finally block, could not be done because that variable is removed earlier.)