Fix SF update app versions bug when package folder doesn't exist (#5090)
When using Docker-based Service Fabric service projects, there will not exist a code package within the service package generated by the build. This is simply because a code package is not relevant; the code exists in the Docker image. In such a scenario, however, the ServiceFabricUpdateAppVersions task fails its validation logic because it assumes a Code folder will exist in the service package since the service manifest described a Code package. This leads to an error when running that task.
The solution is to change the logic in the task to account for the situation where the folder does not exist. If it doesn't, then it simply skips further analysis of that folder.