Changed input to hidden jenkins project type.
if the value does not set already, (assume jenkins offline server case) it will try to talk to server to figure out the type.
Based on the type it will decide the if its multibranch pipeline or not.

If the server is reachable, the input is already auto populated and subsequent inputs will based on that fetch the values.
Changed commits file suffix input hidden as suggested.