Changed Remove-Item logic for temp log folder.
Errors in RemoteDeployer when run as part of the PowerShellOnTargetMachineV3 task. Could not run more than one task in the release pipeline without getting the "Remove-Item : Windows PowerShell is in NonInteractive mode. Read and Prompt functionality is not available." error. This change changes Remove-Item to work as in a similar issue listed here: https://stackoverflow.com/questions/34632351/remove-item-errors-when-run-at-end-of-tfs-build. Have confirmed I can now run multiple PowerShellOnTargetMachineV3 tasks in sequence.