Chaining async execution of unzipExtract to run.
The Java Tool Installer Task fails when the jdk is provided as a zip
file.

The issue is caused by the extractFiles method not awaiting the
async result of the unzipExtract method. Instead the extractFiles method
returns before the unzip powershell command has been executed causing the task to fail.

Adressed the issue by chaining the async execution of the unzipExtract
method through to the async run method of the task.