FIX WHERE CASED .CONFIG FILES ARE IGNORED (#7583)
Minor fix to correctly strip away the .config suffix for files during
the search for the appropriate environment config file. Without the fix,
for an XML file labelled Web.Config (note the uppercase C), the script would
try and search for a transform file of Web.Config.Dev.config for an
environment named 'Dev'.  This patch addresses the problem.