Tag Filters: GitHub Release Task (#10703)
Enhancements for tags: User defines the tag pattern for which the release should be created.

When tagSource is auto and tagPattern is specified,

In case of manual trigger, Find a tag matching the tagPattern at the target. If there are multiple matching tags or no matching tags, Do not create a release.
If the trigger is a commit push, Find a tag on that commit matching the tag Patter, If there are multiple matching tags or no matching tags, Do not create a release.
If the trigger is a tag push, Check if that tag matches the given pattern. If yes, create a release, else do not.
Adding a new optional parameter tagPattern to getTagForCommitTarget and _getTagForCommit.
If the tagPattern is specified, tags are filtered based on the tagPattern too in addition to the existing filters.