Use v4 UUID instead of v1 for test feedId
This uses a purely random v4 UUID instead of a time-based v1 UUID for
the feedId in the nuget test (which was introduced in
c265d2a6feb9773b15e5c15ac51480cdd9b00da2).

v1 UUID are based on current time and the hardware MAC address of the
machine where they are being generated (although the implementation in
the npm uuid module uses generates a random fake MAC address). As such
they have much more complex semantics than v4 UUIDs which are simply
randomly generated.

Unless there's a specific requirement for the special semantics of v1
UUIDs it is simpler and less error prone to simply go for v4 UUIDs
whenever just a unique identifier is needed.