Changed JS replace function from string to regex version where needed (#12508)
* Changed single-time javascript replacements (string replacement), which should actually be global replacements, to global replacements (regex replacement).

This is due to a quirks in javascript which is often overlooked: pure string replacements will only replace the FIRST occurrence of the string. However, most of the time you want to replace all occurences. For this you need to use the regex version of the javascript replace function. So I changed the string-replacements to regex-replacements where every occurence of a string should be replaced.

* Bumped version numbers of changed tasks

Co-authored-by: Szumovski Markus <m.szumovski@pke.at>