Fix for unzipping archives with lot of files (#13958) (#13995)
Issue: If any option File Transforms or Variable Substitution is enabled, the zip package is extracted to perform the selected operation when source type is zip. The task uses decompress-zip node module to perform the extraction and for packages with lot of files, the module fails with EMFILE error.
Fix: Instead of using the decompress-zip module for extraction, changed it to consume tools 7zip for windows and unzip for non-windows agents.