AppCenterDistributeV3: Use newer version of the "appcenter-file-upload-client-node" (#15604)
AppCenterDistributeV3

Increase the version used of package "appcenter-file-upload-client-node" to version that already has content-type for aab. This is a needed fix for aab distribution in AppCenter.
Also add "skipLibCheck" to our task's tsconfig, as that's the setup that we also use in appcenter-cli and without it, we would be getting TS errors for our internal appcenter-file-upload-client-node package, probably because of many conflicting package versions, such as of TS, node, npm, ... between the packages.