Fix Python Twine Upload Authenticate Task
The Python Twine Upload Authenticate task will fail on feeds that contain a dot in their
name. The dot in the feed name will be escaped in the generated .pypirc
file in the corresponding section headers (e.g. [Feed\.Name]).

The issue is due to how the `ini` npm package encode works. Thus
providing a custom `toString` implementation that convert repositories
to their .pypirc representation fixes the issue.

Resolves: #14220