Bump AzureFunctionAppV1 dep azure-storage to latest
This latest version is deprecated, but it is latest, so as a stepping stone to migrating to its successor, we're bumping the version now. There is no automated test yet, I will do a manual smoke test against a private instance of ADO and attempt to add an automated test later.