Remove the rule enforcing package.json can only have dependencies and not devDependencies
At some point this was introduced. We are reaching out to see why, but we are fairly sure we want to remove this. The reason why is that if we keep the practice of having package.json to bump when the task code changes and Tests/package.json to bump when the test code changes, test code changes won't trigger NPM releases as the task code is unaffected by those. This means that tasks which rely on dependencies from azure-pipelines-tasks on NPM will not see the updated task if only the Tests.package.json file changes and will continue to get reported by the component governance check. If we want a test-only change to trigger a deployment, we need to change the root package.json also, at which point the two packages are not independent anyway. So there is no value in keeping two package.json files.