Bug on windows machines in ci/verifyMinAgentDemands execution via pipeline (#3087)
When you access another batch file from a batch file, you need to use the CALL command to return control to the parent process.
Otherwise control is passed to the batch file being executed. So CmdLineV2 Task was replaced with 2 separated commands
Because npm install on windows is a batch file which call npm-cli.js and you cant use "call npm install" command on Linux-based machines