Add experimental feature to AzureCLIV2 which keeps the session running in the background in the case of ARM service connections with workload identity federation scheme (#19989)
Today, any attempts made to get the access token for a different resource than originally will fail with the Entra error AADSTS700024 after the IdToken expires, which is after 10 minutes. The same happens 60 minutes after the issuance date of the original access token, when it expires, and we cannot refresh. This change is a tactical solution to this problem that is to be used until az-cli supports this internally. This is an EXPERIMENTAL feature which means we provide no guarantees and no support for it. That said, it should work well for 99.9% of the cases. Please note that this feature can be removed at any point in time, particularly when az-cli delivers the long-term solution to this problem.