[AzureCLI] Fix DNS lookup order for node20
Started from the node v17, dns lookup order changed from the ipv4first to verbatim.
Because of that the callback receives IPv4 and IPv6 addresses in the order the DNS resolver returned them.
In some cases when we try to aquire token, the ipv6 addres is unreachable which cause ENETUNREACH and ETIMEDOUT error after that.
Since the problem is only with node20, restore the dns lookup order to ipv4first only for node20.