Update CoreAz.ps1 to use the latest Resolve-AzError
The script is using Resolve-Error old commandlet from 12.4.0 but this commandlet is deprecated in newer Azure Powershell versions and is failing for Self-hosted agents with newer version of Azure Powershell. So customers are having to downgrade Azure Powershell version to make this task work.


##[error]Process 'openssl.exe' exited with code '-1073741515'.
An error occurred in Initialize-AzModule
##[error]The term 'Resolve-Error' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again.
##[error]PowerShell exited with code '1'.;
..
..
System.Management.Automation.CommandNotFoundException: The term 'Resolve-Error' is not recognized as the name of a cmdlet, function, script file, or operable program. Check the spelling of the name, or if a path was included, verify that the path is correct and try again.